Ajout arrondi au dixième pour factures reçues
- Nouvelle option 'Dixièmes (0.1)' dans l'arrondi

- Changement parseInt → parseFloat pour supporter 0.1

- Ex: 1234.67 → 1234.7

+++ app.js
@@ -2317,7 +2317,7 @@ function calculateRecInvVat() {
 
     // Apply rounding if selected
     if (rounding !== 'none') {
-        const roundTo = parseInt(rounding);
+        const roundTo = parseFloat(rounding);
         total = Math.round(total / roundTo) * roundTo;
         // Recalculate VAT to match rounded total
         vat = total - subtotal;
